pyproject.toml
[project]
name = "modular-agent-designer"
version = "0.1.0"
description = "A modular framework for designing and orchestrating complex agentic workflows with ease."
requires-python = ">=3.13"
dependencies = [
    "google-adk[extensions]==2.0.0b1",
    "pydantic>=2,<3",
    "pyyaml>=6",
    "httpx>=0.27",
    "click>=8",
    "flake8>=7.3.0",
]

[project.scripts]
modular-agent-designer = "modular_agent_designer.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
]
telemetry = [
    "mlflow>=2.14.0",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp-proto-http",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/modular_agent_designer"]

[tool.uv]
prerelease = "allow"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
